fix(api): make applicant_id a required field for making applications require applicant_id for association of applications to applicant

+++ api/app/controllers/api/v1/applications_controller.rb
@@ -16,7 +16,9 @@ module Api::V1
 
         # POST /applications
         def create
+            params.require(:applicant_id)
             if invalid_id(Session, :session_id) then return end
+            if invalid_id(Applicant, :applicant_id) then return end
             application = Application.new(application_params)
             if not application.save # does not pass Application model validation
                 application.destroy!
@@ -90,7 +92,6 @@ module Api::V1
         end
 
         def valid_applicant_matching_data(errors)
-            if invalid_id(Applicant, :applicant_id) then return end
             matching = ApplicantDataForMatching.new(applicant_data_for_matching_params)
             if matching.save
                 return nil
